+++ CHANGELOG.md
@@ -4,8 +4,27 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
-### Bug Fixes
+### 🐛 Bug Fixes
 
 - Inspection page jumping during updates
+- Fronted version
+
+### 🛠️ Engineering
+
+- Start using CHANGELOG.md with tools added
+
+
+
+## [0.2.0] - 2026-03-20
+
+- Support Google Gemini backend
+- Fix some bugs
+
+
+## [0.1.1] - 2026-3-13
+
+- Improve UI and UX
+- Fix some bugs
+- Add some unit tests